[#82] fix: CoreData 더미 데이터 생성 코드 분리
